refactor: 애니메이션 계산 로직 분리
